+++ packages/mg-components/src/components/atoms/mg-input-title/test/mg-input-title.spec.tsx
@@ -9,14 +9,13 @@ const getPage = args =>
   });
 
 describe('mg-input-title', () => {
-  test.each([
-    { required: false, identifier: 'identifier' },
-    { required: true, identifier: 'identifier' },
-    { isLegend: true, identifier: 'identifier' },
-  ])('Should render label with args %s', async args => {
-    const { root } = await getPage(args);
-    expect(root).toMatchSnapshot();
-  });
+  test.each([true, false].flatMap(isLegend => [true, false].flatMap(required => [true, false].map(readonly => ({ isLegend, required, readonly, identifier: 'identifier' })))))(
+    'Should render label with args %s',
+    async args => {
+      const { root } = await getPage(args);
+      expect(root).toMatchSnapshot();
+    },
+  );
 
   test.each(['', ' ', undefined])('Should throw error, case invalid identifier prop', async identifier => {
     expect.assertions(1);