+++ packages/mg-components/src/components/molecules/inputs/mg-input-checkbox/test/mg-input-checkbox.spec.tsx
@@ -103,10 +103,13 @@ describe('mg-input-checkbox', () => {
       }
     });
 
-    test('Should not render with invalid value property: %s', async () => {
+    test.each([
+      { identifier: 'identifier', type, label: 'label', value: undefined },
+      { identifier: 'identifier', type, label: 'label', value: ['batman', 'joker', 'bane'] },
+    ])('Should not render with invalid value property: %s', async args => {
       expect.assertions(1);
       try {
-        await getPage({ identifier: 'identifier', type, label: 'label', value: ['batman', 'joker', 'bane'] });
+        await getPage(args);
       } catch (err) {
         expect(err.message).toMatch('<mg-input-checkbox> prop "value" is required and all values must be the same type, CheckboxItem.');
       }