+++ packages/mg-components/src/components/molecules/inputs/mg-input-text/mg-input-text.conf.ts
@@ -1,17 +1,8 @@
 /**
- * List of all availables text type
- */
-export const textTypes = ['text', 'search'] as const;
-
-/**
  * TextType from text types
  */
-export type TextType = (typeof textTypes)[number];
 
-/**
- * Type for datalist option
- */
-export type DatalistOption = string;
+export type TextType = 'text' | 'search';
 
 /**
  * Available handlers